Refactored UndoAction and RedoAction into combined abstract action class
