+++ JHotDraw/.idea/libraries/Maven__dk_sdu_mmmi_featuretracer_1_0.xml
@@ -1,0 +1,13 @@
+<component name="libraryTable">
+  <library name="Maven: dk.sdu.mmmi:featuretracer:1.0">
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/dk/sdu/mmmi/featuretracer/1.0/featuretracer-1.0.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/dk/sdu/mmmi/featuretracer/1.0/featuretracer-1.0-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/dk/sdu/mmmi/featuretracer/1.0/featuretracer-1.0-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
