+++ JHotDraw/src/main/java/org/jhotdraw/app/action/RedoAction.java
@@ -1,106 +1,28 @@
-/*
- * @(#)RedoAction.java  2.0  2006-06-15
- *
- * Copyright (c) 1996-2007 by the original authors of JHotDraw
- * and all its contributors.
- * All rights reserved.
- *
- * The copyright of this software is owned by the authors and  
- * contributors of the JHotDraw project ("the copyright holders").  
- * You may not use, copy or modify this software, except in  
- * accordance with the license agreement you entered into with  
- * the copyright holders. For details see accompanying license terms. 
- */
-
 package org.jhotdraw.app.action;
-import dk.sdu.mmmi.featuretracer.lib.FeatureEntryPoint;
-import java.awt.*;
-import java.awt.event.*;
-import javax.swing.*;
-import javax.swing.text.*;
-import java.beans.*;
-import java.util.*;
 import org.jhotdraw.util.*;
 import org.jhotdraw.app.Application;
-import org.jhotdraw.app.JHotDrawFeatures;
-import org.jhotdraw.app.View;
 /**
  * Redoes the last user action.
  * In order to work, this action requires that the View returns a project
  * specific undo action when invoking getAction("redo") on the View.
- *
- *
+ * 
+ * Is the specifc implementation for redoing, extending AUndoRedoAction. 
+ * 
  * @author Werner Randelshofer
- * @version 2.0 2006-06-15 Reworked.
+ * @version 6.9 2020 Reworked.
  * <br>1.0 October 9, 2005 Created.
  */
-public class RedoAction extends AbstractViewAction {
+public class RedoAction extends AUndoRedoAction {
     public final static String ID = "edit.redo";
-    private ResourceBundleUtil labels = ResourceBundleUtil.getBundle("org.jhotdraw.app.Labels");
     
-    private PropertyChangeListener redoActionPropertyListener = new PropertyChangeListener() {
-        public void propertyChange(PropertyChangeEvent evt) {
-            String name = evt.getPropertyName();
-            if (name == AbstractAction.NAME) {
-                putValue(AbstractAction.NAME, evt.getNewValue());
-            } else if (name == "enabled") {
-                updateEnabledState();
-            }
-        }
-    };
+    //Cannot be declared in super-class.
+    private ResourceBundleUtil labels = ResourceBundleUtil.getBundle("org.jhotdraw.app.Labels");
+ 
     
     /** Creates a new instance. */
     public RedoAction(Application app) {
         super(app);
+        setID(ID);
         labels.configureAction(this, ID);
     }
-    
-    protected void updateEnabledState() {
-        boolean isEnabled = false;
-        Action realRedoAction = getRealRedoAction();
-        if (realRedoAction != null) {
-            isEnabled = realRedoAction.isEnabled();
-        }
-        setEnabled(isEnabled);
-    }
-    
-    @Override protected void updateView(View oldValue, View newValue) {
-        super.updateView(oldValue, newValue);
-        if (newValue != null && newValue.getAction(ID) !=  null) {
-            putValue(AbstractAction.NAME, newValue.getAction(ID).
-                    getValue(AbstractAction.NAME));
-            updateEnabledState();
-        }
-    }
-    /**
-     * Installs listeners on the view object.
-     */
-    @Override protected void installViewListeners(View p) {
-        super.installViewListeners(p);
-        if (p.getAction(ID) != null) {
-        p.getAction(ID).addPropertyChangeListener(redoActionPropertyListener);
-        }
-    }
-    /**
-     * Installs listeners on the view object.
-     */
-    @Override protected void uninstallViewListeners(View p) {
-        super.uninstallViewListeners(p);
-        if (p.getAction(ID) != null) {
-        p.getAction(ID).removePropertyChangeListener(redoActionPropertyListener);
-        }
-    }
-
-    @FeatureEntryPoint(JHotDrawFeatures.UNDO_REDO)
-    public void actionPerformed(ActionEvent e) {
-        Action realRedoAction = getRealRedoAction();
-        if (realRedoAction != null) {
-            realRedoAction.actionPerformed(e);
-        }
-    }
-    
-    private Action getRealRedoAction() {
-        return (getActiveView() == null) ? null : getActiveView().getAction(ID);
-    }
-    
 }
