+++ JHotDraw/src/main/java/org/jhotdraw/undo/CompositeEdit.java
@@ -15,7 +15,6 @@
 package org.jhotdraw.undo;
 
 import javax.swing.undo.*;
-import java.util.*;
 /**
  * This is basically the same like javax.swing.undo.CompoundEdit but
  * it has a slightly different behaviour:
